requirements.txt
pytest~=8.3.4
selene==2.0.0rc7
allure-pytest==2.13.2